Corrijo bugs del último merge. Refactoreo
