+++ Ta-C-Ti.depend
@@ -144,24 +144,23 @@
 	<stdlib.h>
 	"secundarias.h"
 
-1749776577 source:e:\documentos\unlam win\2025\ayed\ta-c-ti\api.c
+1749862900 source:e:\documentos\unlam win\2025\ayed\ta-c-ti\api.c
 	<stdio.h>
 	<stdlib.h>
 	<string.h>
 	"ranking.h"
-	"jugadores.h"
 	"api.h"
 
-1749776577 e:\documentos\unlam win\2025\ayed\ta-c-ti\ranking.h
+1749862890 e:\documentos\unlam win\2025\ayed\ta-c-ti\ranking.h
 	"lista.h"
-	"jugadores.h"
+	"main.h"
 
 1749776577 e:\documentos\unlam win\2025\ayed\ta-c-ti\lista.h
 	<stdio.h>
 	<string.h>
 	<stdlib.h>
 
-1749777654 e:\documentos\unlam win\2025\ayed\ta-c-ti\jugadores.h
+1749860412 e:\documentos\unlam win\2025\ayed\ta-c-ti\jugadores.h
 
 1749776577 e:\documentos\unlam win\2025\ayed\ta-c-ti\api.h
 	"lista.h"
@@ -174,7 +173,7 @@
 	"stdio.h"
 	"stdlib.h"
 
-1749777615 source:e:\documentos\unlam win\2025\ayed\ta-c-ti\juego.c
+1749861935 source:e:\documentos\unlam win\2025\ayed\ta-c-ti\juego.c
 	<stdio.h>
 	<stdlib.h>
 	"secundarias.h"
@@ -182,17 +181,17 @@
 
 1749776570 e:\documentos\unlam win\2025\ayed\ta-c-ti\secundarias.h
 
-1749777632 e:\documentos\unlam win\2025\ayed\ta-c-ti\juego.h
+1749861943 e:\documentos\unlam win\2025\ayed\ta-c-ti\juego.h
 	"main.h"
 	<time.h>
 
-1749777218 e:\documentos\unlam win\2025\ayed\ta-c-ti\main.h
+1749862294 e:\documentos\unlam win\2025\ayed\ta-c-ti\main.h
 
-1749776577 source:e:\documentos\unlam win\2025\ayed\ta-c-ti\lista.c
+1749862924 source:e:\documentos\unlam win\2025\ayed\ta-c-ti\lista.c
 	"Lista.h"
-	"jugadores.h"
+	"main.h"
 
-1749776577 source:e:\documentos\unlam win\2025\ayed\ta-c-ti\ranking.c
+1749862294 source:e:\documentos\unlam win\2025\ayed\ta-c-ti\ranking.c
 	<stdio.h>
 	<string.h>
 	<time.h>
@@ -212,13 +211,15 @@
 	<stdlib.h>
 	"secundarias.h"
 
-1749776870 source:e:\documentos\unlam win\2025\ayed\ta-c-ti\main.c
+1749863026 source:e:\documentos\unlam win\2025\ayed\ta-c-ti\main.c
 	<stdio.h>
 	<time.h>
 	"main.h"
 	"secundarias.h"
 	"ranking.h"
+	"cola.h"
 	"api.h"
 	"lista.h"
-	"jugadores.h"
+	"RivalIA.h"
+	"Juego.h"
