+++ api_response.json
@@ -1,1 +1,1 @@
-[{"nombreJugador":"Fede","puntaje":69,"ultimaPartidaJugada":"12/06/2025 18:17:18"},{"nombreJugador":"Delfi","puntaje":68,"ultimaPartidaJugada":"12/06/2025 18:12:05"},{"nombreJugador":"Pedro","puntaje":2,"ultimaPartidaJugada":"12/06/2025 18:02:32"},{"nombreJugador":"Juani","puntaje":3,"ultimaPartidaJugada":"12/06/2025 18:17:18"},{"nombreJugador":"Elmo","puntaje":9,"ultimaPartidaJugada":"12/06/2025 20:42:21"},{"nombreJugador":"Moe","puntaje":8,"ultimaPartidaJugada":"12/06/2025 21:44:26"},{"nombreJugador":"Jorge","puntaje":5,"ultimaPartidaJugada":"12/06/2025 22:09:40"},{"nombreJugador":"Kris","puntaje":-3,"ultimaPartidaJugada":"12/06/2025 22:30:49"},{"nombreJugador":"Susie","puntaje":3,"ultimaPartidaJugada":"12/06/2025 22:17:26"},{"nombreJugador":"Fa","puntaje":7,"ultimaPartidaJugada":"12/06/2025 22:32:16"}]
+[{"nombreJugador":"Fede","puntaje":69,"ultimaPartidaJugada":"12/06/2025 18:17:18"},{"nombreJugador":"Delfi","puntaje":68,"ultimaPartidaJugada":"12/06/2025 18:12:05"},{"nombreJugador":"Pedro","puntaje":2,"ultimaPartidaJugada":"12/06/2025 18:02:32"},{"nombreJugador":"Juani","puntaje":3,"ultimaPartidaJugada":"12/06/2025 18:17:18"},{"nombreJugador":"Elmo","puntaje":9,"ultimaPartidaJugada":"12/06/2025 20:42:21"},{"nombreJugador":"Moe","puntaje":8,"ultimaPartidaJugada":"12/06/2025 21:44:26"},{"nombreJugador":"Jorge","puntaje":5,"ultimaPartidaJugada":"12/06/2025 22:09:40"},{"nombreJugador":"Kris","puntaje":-3,"ultimaPartidaJugada":"12/06/2025 22:30:49"},{"nombreJugador":"Susie","puntaje":3,"ultimaPartidaJugada":"12/06/2025 22:17:26"},{"nombreJugador":"Fa","puntaje":7,"ultimaPartidaJugada":"12/06/2025 22:32:16"},{"nombreJugador":"Facundo","puntaje":11,"ultimaPartidaJugada":"13/06/2025 22:12:52"},{"nombreJugador":"Prueba1306","puntaje":8,"ultimaPartidaJugada":"13/06/2025 22:26:36"}]
