--- jugadores.h
@@ -1,16 +1,0 @@
-#ifndef JUGADORES_H_INCLUDED
-#define JUGADORES_H_INCLUDED
-
-typedef struct {
-    char nombre[50];
-    int puntaje;
-    int partidasJugadas;
-    int partidasRestantes;
-} Jugador;
-
-void iniciarJuego();
-int partida(char *nombreJug);
-
-#endif // JUGADORES_H_INCLUDED
-
-
