+++ post_body.json
@@ -2,7 +2,7 @@
   "CodigoGrupo": "proceso",
   "Jugadores": [
     {
-      "nombre": "Fa",
+      "nombre": "Prueba1306",
       "puntos": 7
     }
   ]
